use `users: UserConnection!` in graphql API to allow for totalCount field
See https://github.com/sourcegraph/sourcegraph/pull/8550 for the description of and motivation for `XyzConnection`-style GraphQL types.

Also updates all callers. (There are no callers or docs outside this repo.)

Note that this also removes a redundant is-admin check in `(*schemaResolver).Users`. The check is also performed in `(backend.Users).List`.