do not run go profiler debugservers by default
This did not work anyway because multiple processes all tried to use the same default :6060. If you want to run a profiler, you'll need to do so manually (e.g., by setting SRC_PROF_HTTP env var in the Procfile for the process you want). This is what you would have needed to do before this change as well.