decode secret key via JSON config as base64, supply key as []byte
Fixes an issue on org invites where it would fail and report 'key is invalid'.

Also fixes another issue where the secret key value would not be base64-decoded, which would mean that setting the same value in SRC_APP_SECRET_KEY vs the JSON config would result in different keys being used in the invite code generation. This was not a security vulnerability but could lead to invites not being redeemable.