reduce noisy user_activity_user_id_pkey pgsql logs, remove user_activity table
The `user_activity` table was unnecessary because it was 1-to-1 with `users` rows. It being a separate table required complexity to perform upserts (since are still running postgresql 9.4). To remove that complexity, and suppress the errors from failed inserts that were ok at the app level, I just moved the user activity columns to the `users` table.