searcher: Limit concurrent fetches to gitserver (#8786)
This commit adds a semaphore around FetchTar, which is responsible for speaking
to gitserver. This is to avoid attempting to fetch thousands of repos at once,
leading to no repos finishing.

We also add aggressive timeouts on waiting for a fetch to start (200ms). The
frontend already nicely handles timeouts indicating to the user to try again in
the future. The aggressive timeout on store is using the same result field
(timedout).