wait for frontend to be available in repo-updater
The repo-updater and frontend often start up at the same time. When the repo-updater starts up, it immediately tries to contact the frontend. If the frontend is unavailable, it logs an error. This error usually is just temporary and can be ignored. But the error is noisy.

So a fix I added is to just sleep for an arbitrary time period in repo-updater before trying to access the frontend. This "fixes" the log noisiness issue but introduces delay for users setting up a server. This is a better solution.