add setting search.repositoryGroups
Users, orgs, and sites (in the `settings` field of site config) can specify a new setting `search.repositoryGroups` like so:

```
{
  "search.repositoryGroups": {
    "foo": ["github.com/moby/moby","github.com/golang/oauth2"]
  },
  "search.scopes": [
    {"name": "Foo repos","value": "repogroup:foo"}
  ]
}
```

Then users can select the `Foo repos` scope, and it will search only in the repositories listed under the `foo` repogroup.

It does not query the DB using `Repos.List` for each of the repos listed in the `foo` repogroup.