searcher: Only release gitserver semaphore when finished running archive (#8894)
* searcher: Only mark as done when reader is closed

* searcher: Simplify store fetch timeout

Now that we have `done` which is called when the function fails or the stream is
closed, we can consolidate timeout code into that and remove our `afterFunc`
logic.

* searcher: Only release gitserver semaphore when finished running archive

Previously we would release the gitserver semaphore once we returned the
io.ReadCloser. However, that means we would let more `git archives` happen
while other ones are running, effectively making the semaphore very
ineffective. This change ensures we only release the semaphore when we are
done speaking to gitserver for an archive.

* searcher: Instrument fetch queue size and fetches