retroactively fix repo.blocked->repo.enable migration
c6f0adfb43111f8ffb1222d7625cbff9dda067e0 was a bad migration because it would lock on servers that had active Repos.List queries. This fixes the migration. Note that there are 2 possible states for servers whose DB has schema_migrations.version=1515125883 now, and we will need to address that in a future migration.