pkg/endpoint: Return error if we can't List endpoints (#8951)
* pkg/endpoint: Update Get to return error

We want to indicate errors communicating to K8S after New.

* pkg/endpoint: Return error if we can't List endpoints

At Umami we have RBAC enabled, so we failed to list the searcher
endpoints. This listed to us just relying on the searcher service, which lead
to a bad performance bug since we rarely hit the same server for a
repo@commit. This updates our implementation to fail searches if we
misconfigure kubernetes, rather than just have terrible performance.

* pkg/endpoint: Use a simple mutex

We can just hold the mutex before kicking of the initialize goroutine. This
allows us to remove the complicated condition behaviour.