searcher: remove srcFile.zipFile
The back-reference was a vestige of an old design.
This shrinks the zipFile struct.

This makes the zipFile 24 bytes (on 64 bit systems).
The current Go memory allocator doesn't have a 24 byte slab,
so a 32 byte one gets used, which means that this doesn't
have any immediate impact.

However, the next commit switches from []*srcFile to []srcFile,
and with that change, we will actually realize the 25% memory savings.