web: Remove timing stat and correctly bound result count stat (#9009)
* web: Remove timing from search result stats

The timing is not consistent, so can be confusing to users.

* web: Count number of file matches rather than line

The backend limits by file matches, not line matches. If we want to present by
line matches, we should update the backend to limit by line matches. This is to
avoid the number of results statistics to be inconsistent between searches.