pkg/gitserver: select correct gitserver to use for the repo
Previously IsRepoCloned, IsRepoCloneable, and EnqueueRepoUpdate always called the first gitserver even if there were multiple. This would cause them to be somewhat ineffective. Repos would be duplicate-cloned on both the intended gitserver and the first gitserver, and updates would only be enqueued when the intended gitserver was accessed by running a command on it, not by EnqueueRepoUpdate. So the results of this bug weren't catastrophic.