gitserver RepoInfo, consolidate repo mirror info responses in GraphQL type
This adds a new omnibus RepoInfo RPC method to gitserver that returns cloneInProgress,
cloned, updatedAt, and the URL. Adding a new RPC method that returned all of these fields
seemed better than adding individual RPC methods for the two additional new fields because
this way is less code and fewer RPC calls.

These fields are all then grouped in the GraphQL API.

It is also now possible to determine if a repo has been cloned without triggering a clone
of it in the GraphQL API. This will be important in a subsequent commit.