gitserver: remove the ability to /list?cloning (list repos that are cloning in progress) - fix #9085
It is no longer necessary for the UI to know this, and there are no other users of this. I added this recently but found a better UX to help the site admin while setting up their server that does not involve showing them the cloning status of their site's repos.