show search duration (elapsed msec), computed on backend
We took this out when we still had the auto-reload-results-when-any-is-cloning behavior because that meant the duration jumped around a lot, which looked glitchy. Now that we don't have that behavior, it's OK to add it in. It also measures the timing on the backend, which is how Google does it and is a more reliable measurement. This will help when eyeballing search perf.