copy monaco assets on disk when using webpack dev server
This is necessary to use the Monaco settings editor with webpack dev server if you have never run webpack with non-webpack-dev-server. The files must exist on disk.

The alternative would be to use https://github.com/Microsoft/monaco-editor#integrate-cross-domain (because localhost:3088 WDS is on a different host from localhost:3080), but that's more complex and results in less consistency between dev and prod.