store external repo ID and service name/URL for repositories (e.g., pointing to their github repo)
Introduces a more robust way to link a repository on Sourcegraph with its corresponding external (GitHub/Gitolite/etc.) repository. This information is a 3-tuple `(external id, external service type, external service id)` such as:

- for github.com/nodejs/node: `(MDEwOlJlcG9zaXRvcnkyNzE5Mzc3OQ==, github, https://github.com/)`, which is `(github graphql api node id for repository, constant type string "github", base url to github instance the repo lives on)`

This information is provided by the repo-updater for configured GitHub instances and is backfilled in each call to `backend.Repos.GetByURI`. The information is not yet used, but in future commits it'll be used to (in order):

- cleanly support repos from multiple code hosts (without github.com/github-specific code, checking for `"github.com/"` URI prefix, etc.)
- make the DB, not gitserver, the source of truth for a repo's clone URL
- show rich code host-specific info about a repo (such as stars, followers, PR-annotated branch list, searchable issues/PRs, etc.)
- robustly handle renames and deletions of external repositories
- set up code host-specific integrations (such as GitHub webhooks to get fresher code)