diff search: adjust individual ops' timeouts to finish (even partially) before parent timeout (#9213)
When diff search was changed (by me) to exec `git log` then `git show` (which is required for reasons described in the code file), its ability to return partial results if the context timed out was broken. That's because it would run the first `git log` command until the context timed out (naively), so the second `git show` command would never have time to return anything, thus the whole return value would be empty.

This uses a heuristic of letting each of those 2 ops run for 48% of the parent context time.

Note: There is significant room for further perf improvement here, of course. The two operations duplicate a TON of effort, we could parallelize `git show`, etc.).