gitserver: accept Git remote URL (along with repo URI) in requests (#9206)
* gitserver: accept Git remote URL (along with repo URI) in requests

This is the first step in moving away from gitserver or originmap being the source of truth for Git remote URLs. It makes gitserver use the Git remote URL specified in the RPC, if the field is set. If not, it looks up the URL using originmap or `git remote get-url ...` (so it's 100% backcompat with current gitserver API clients).

* review comment: use context

* gitserver: remove docstrings about removing Repo (it will stay)

* fall back to originmap if it exists, to reduce impact of change