all: fix multiple HTTP connection / goroutine leaks (#9234)
* all: fix multiple HTTP connection / goroutine leaks

This change fixes a few HTTP connection / goroutine leaks, where we were
failing to close the response body of HTTP requests.

So far as I can tell, this affected:

- `InternalClient.RetryPingUntilAvailable`
- `InternalClient.ReposCreateIfNotExists`
- `InternalClient.ReposUnindexedDependencies`
- `InternalClient.ReposGetByURI`
- `InternalClient.ReposGetInventoryUncached`

In addition to the endpoints I added / code I duplicated for cmd/query-runner. If any other services invoke the above methods prior to this change, they would leak one goroutine per call as far as I can tell. This is because Go wouldn't release the internal goroutine, see https://stackoverflow.com/a/45860425/5905344

* pkg/api: Move marshalling to post helper function

This allows us to correctly close the response body without any
leaking. Previously there would still be several cases we would leak the
body. Also introduces using the context, so that we can cancel requests when
our context is done.

* pkg/api: Use POST for ReposGetByURI

The handler doesn't care, so we might as well use post like every other method
does. We get the nice extra logic if we do this.

* pkg/api: Use context when pinging