vendor: use latest github.com/mattes/migrate to fix psql "can't acquire lock" error in db tests (#9236)
In the switch to dep, our version of github.com/mattes/migrate was downgraded such that
the following error always occurred for me locally and on CI:

```
$ go test ./cmd/frontend/internal/db -run= -v
=== RUN   TestMigrations
--- PASS: TestMigrations (4.07s)
=== RUN   TestErrorsInterface
--- PASS: TestErrorsInterface (0.00s)
=== RUN   TestGlobalDeps_TotalRefsExpansion
--- PASS: TestGlobalDeps_TotalRefsExpansion (0.00s)
=== RUN   TestGlobalDeps_update_delete
--- PASS: TestGlobalDeps_update_delete (4.07s)
	global_deps_test.go:105: update
	global_deps_test.go:124: delete other
	global_deps_test.go:142: delete
=== RUN   TestGlobalDeps_RefreshIndex
2018/01/30 04:42:01 can't acquire lock
FAIL	sourcegraph.com/sourcegraph/sourcegraph/cmd/frontend/internal/db	11.001s
```

This issue was fixed by an upstream change that runs migrations in the same conn instead of
from any conn in the pool. This is important because the migrate lib uses pg advisory
locks, which must be unlocked by the same conn (session) that locked them in the first place.