gitserver: set dir when looking up repo's remote during update (#9238)
* gitserver: set dir when looking up repo's remote during update

This is a bad bug in my #9206 PR that is likely to have caused alarming behavior on localhost (all repos get replaced with sourcegraph/sourcegraph locally, because you're probably running `dev/start.sh` in sourcegraph/sourcegraph) and that halted repo updating on prod and dogfood (but did not cause repos to be replaced with the wrong repo, because the cwd of the process is not in a git repo when deployed).

Fixes #9232

* gitserver: improve backcompat for clients that do not send repo remote url

If clients do not send the repo remote URL in their gitserver request (which no clients do yet), this makes it look up the URL quickly in the originmap by default and does not log (to reduce log noise).

This is just nicer backcompat behavior, esp. since there are no clients that have the new behavior yet.