frontend: remove duplicate repo get-by-uri route registration (#9239)
This `repos/get-by-uri` URL path was not being used anywhere. It appears to have caused a bug where the actually used route (also registered with the same name `ReposGetByURI` right below) was not being properly routed to. This caused the indexer to be unable to index repositories, reporting:

```
05:36:35           indexer | t=2018-01-30T05:36:35-0800 lvl=eror msg="Indexing failed" repo=github.com/kubernetes/kops rev= error="internal API response error code 405 (http://localhost:3090/.internal/repos/github.com/kubernetes/kops)"
```

(405 is Method Not Allowed, which makes sense.)

My guess is that the switch to dep changed the gorilla/mux version we use, and the two versions handle this case differently. Perhaps the version we were on before applied the POST method from the 1st route definition to the 2nd route definition, which would have made it work.

(Also, the only consumer of this API is the internal API client, which POSTs to /repos/URI, so I changed the route definition to support only that (not also GET).)