frontend: make redis being unavailable a warning, not an error
Now, if Redis is not reachable on frontend startup, that's OK. It will continue attempting to connect. This is good practice anyway in case there is a Redis blip during when frontend starts up, or if we're debugging a Redis issue causing downtime and need frontend to continue to run.

Confirmed that if Redis is unreachable at startup but later becomes reachable, it establishes the connection and uses Redis from then on.

Also fixes #9312