cmd/query-runner: remove log spam about email notifications not being configured
This warning is really only useful when running with a pre-written site
configuration. For example, in dev mode, or on our sourcegraph.com kubernetes
deployment. But for real users, such as ones running Server via `docker run`,
they must first start the server, see this warning (which is useless to them),
and then visit the site config and configure email sending.

Instead, we just try to send emails regardless and if it fails we will log an
error about emails not be configured.