graphql: Log requests which only fail strict validation (#9347)
* graphql-go: Vendor in latest master

The next commit will modify our manually vendored dir such that its less
strict.

* graphql: Log requests which only fail strict validation

This is a temporary change to allow us to identify bad queries we receive. Once
this stops logging we can switch to the newest version of graphql-go.

* graphql-go: Also log other validation fails

This would be interesting to know. Especially if the new version has other ways
it fails.