db: Use faster password hash for tests (#9351)
* db: Use faster password hash for tests

Currently the db pkg tests are the slowest by far. It isn't due to them
interacting with SQL, its actually due to running bcrypt a bunch. We don't
actually need the security, just correctness for tests. An example of this
change is `TestUsers_BuiltinAuth` goes from taking 12s to 3s on my laptop.

* db: Add test to ensure bcrypt usage is correct