web: disable pointer events on our tooltips (prevents glitching/flickering)
Prior to this change when we would display tooltips it was possible that the
tooltip would in a location overlapping the cursor by just 1px. In such a case,
the tooltip would receive focus and take focus away from the element being
hovered over. The tooltip would then disappear, and come back again once the
cursor became focused on the underlying element again. This would happen quite
quickly as just a quick flashing behavior:

https://www.useloom.com/share/b0b58db9eb3e422280ef98cdea6d1f2a

This fixes that issue by ensuring our tooltips cannot become the target of
mouse events.

Fixes #9358