rcache: Remove potential inifite loop in sysreq (#9588)
If the first time we succeed to connect to redis is after the deadline, then
until redis fails the loop will never terminate. Instead we change the
condition such that if ping ever succeeds we return success, rather than
checking timeouts. The deadline condition is moved into the for loop condition
to make it clear when it will terminate.

We also move the sleep to after the check, so that the happy case doesn't need
to sleep at all.