db: Remove Repos.GetURI (#9754)
This was added to speed up Repo ID -> URI when speaking to the gitserver
backend, due to poor layering where we would only have the ID at the client
layer. We now have the full Repo struct so do not need this call anymore. This
also allows us to remove the repoURICache since it was only read in GetURI.