gitserver: Kill git clones on shutdown (#9819)
Currently when shutting down gitserver any subprocess is just orphaned and
will continue to run until completion (not even respecting context
timeouts). This manifested in a bad way where a customer was experiencing
gitserver taking down there codehost. So they lower the git clone rate
limit. However, this did not take effect since we just restarted the process,
but all the clone processes continued to run (but orphaned from the older
gitserver process).

This commit replaces every use of `context.Background()` with a call to
`contextWithTimeout`. This context is canceled when gitserver receives a
signal to shutdown. The cancel function return additionally also has an
associated `sync.WaitGroup` so we can ensure all uses of `contextWithTimeout`
finish (ie we kill the subprocess).