gitserver: Rate limit ls-remote when cloning. (#9818)
This is a possible root cause of hitting a remote a lot. EnqueueRepoUpdate spins
up cloneRepo in a goroutine, so for example if you have 300 `repos.list` we will
do 300 ls-remote concurrently.

We do not rate limit `IsRepoCloneable` since it blocks a user request. Note:
That request is rare and is only done for admins, so it is safe to not limit in
that case.