search: support type:ref and ref:/hint: for LSP-based references searches (experimental) (#9868)
For example, the following query would find references to any symbol defined in the Go stdlib reflect package in `myrepo`:

  repo:myrepo type:ref lang:go ref:{"package":"reflect"} hints:{"dirs":["file:///"]}

The `ref:` and `hints:` values are treated opaquely by the search backend and come from the LSP `workspace/x*` method results.

This feature is experimental for now and requires opt-in by including a `ref:` operator in your search.

There is no UI support for this feature yet. You need to manually construct the ref and hints JSON values. It is not intended to be actually useful for users yet, just to support the construction of URLs that show references from a certain repo (which is not possible in the current UI).