fix Go symbol URL offset off-by-one error (#9952)
fix #9624

A URL of the form http://localhost:3080/-/godoc/refs?def=New&pkg=github.com%2Fdgrijalva%2Fjwt-go&repo=github.com%2Fdgrijalva%2Fjwt-go should go to the `New` func ident token with the references panel open. It was going to the character AFTER the end of the token, which is incorrect and meant the references panel would not open.