graphqlbackend: expose better APIs for LSP dependencies and packages (#9899)
Greatly enhances the API capabilities for querying information about dependencies and packages (from LSP lang servers with the workspace/xpackages and workspace/xdependencies extensions).

This is used for an upcoming experimental feature in the UI that basically just displays this info. It is also a type of data that we hear customers frequently asking for access to, currently just for experimentation.

Performing reference searches based on dependencies and packages only really works for Go now because this change only includes lang_kludges.go changes for Go. That limitation is acceptable for now and is implicitly documented by the GraphQL API docs saying these fields are experimental.