schema: make maxReposToSearch "unlimited" value -1 instead of 0
- `"maxReposToSearch": -1` BEFORE this change was a silent error, now means "unlimited"
- `"maxReposToSearch": 0` BEFORE this change meant "unlimited", now means "use default (500)".
- `"maxReposToSearch"` not set means "use default (500)" (no change)

This indirectly fixes an issue that Lyft ran into:

> I tried with a string value "maxReposToSearch": "0" but that didn’t seem to do anything [...]

The problem was that because all of our configuration options, including
`maxReposToSearch`, have a generated Go struct with `,omitempty` set in the
`"json:""` struct field tag, when a user explicitly set the value to be `"0"`
(which gets translated to `"maxReposToSearch": 0` by our backwards
compatibility support code), our `,omitempty` causes it to not be marshaled.
This is problematic because `sourcegraph-server-gen` internally marshals and
unmarshals the JSON configuration to turn it into a k8s yaml configmap.

In other words, it is 100% impossible for Data Center users to set a zero
value (`""` for string, 0 for number) for a configuration option and have it be
explicitly set where the frontend can distinguish between
"explicitly set zero value" and just "zero value (because the JSON option was not set)".

Rather than making the above possible, I chose to change `maxReposToSearch` so
that we do not rely on such a distinction. In general, it is very bad to have
the zero value for a configuration option be different than the "default" value
because (1) it is confusing from a user perspective IMHO and (2) issues like
this can easily arise.

Note: Due to the bug, `"maxReposToSearch": 0` AND `"maxReposToSearch": "0"`
previously meant "use default (500)" for ALL Data Center users. Only Server
users will be affected by this change, and only if they had set
`"maxReposToSearch": 0` (they now need to set it to `-1`). My guess is we have
no users who actually did this.

Helps sourcegraph/sourcegraph#9968