pkg/trace: Simplify API and use OT and net/trace combined (#10060)
* Rename pkg/traceutil to pkg/trace

* all: Use new helper pkg/trace.New

This removes the unneccessary extra steps we had before. It also opens up the
possibility of doing these traces with OT as well.

* Move lightstep init into tracer pkg

* trace: Remove race condition on names

I noticed this while refactoring in this PR. If we have more than one goroutine
using this slice and the cap is bigger than len, then both goroutines will write
to the same underlying array. This commit changes us to a simpler model where
instead we just store the prefix for the family name, where the prefix is the
parent family + " > "

* trace: Transparently use both net/trace and OT

There are some areas where we only use net/trace. This will allow us to still
capture these areas in OpenTracing without any changes to our callsites.

* trace: Pass error to SetError

This allows us to simplify callsites of SetError.