conf: require auth.provider to be explicitly set per documentation (#10070)
Currently, if `auth.provider` is set to `"saml"` but `auth.openIDConnect` is set, OpenID Connect auth will be used. There will be warning log about this, but it's still not good behavior.

This change makes it stricter about `auth.provider`. The http-header, saml, and openidconnect auth configs will only be consulted if `auth.provider` is set appropriately.