relax user_emails constraint to prevent users from denying other users access (#10102)
There is a possible "DoS" attack where Mallory, the malicious attacker, adds the email alice@example.com to Mallory's own account. Mallory can't verify this email because she doesn't have access to it. Even though it's unverified, the DB unique constraint would prevent the real Alice from signing up and adding alice@example.com as her email.

This commit relaxes the DB unique constraint so that it:

- Prevents multiple users from having the same verified email address.
- Prevents a user from having duplicates of the same email address associated with their own account.

Now, if Mallory adds alice@example.com to her account and it's unverified, it won't prevent Alice from also adding it.

The DB migration is safe to run because it makes the DB constraint strictly more lenient, so up-migrating the DB is guaranteed to succeed.