{graphql,}backend: remove repo backfill
This was used to set the GitHub ID for github.com repos on sourcegraph.com when we added the `repo.external_*` DB columns. It's been a couple months, and all rows have been backfilled (except for a very small number of very infrequently accessed repositories). So it's safe to remove the backfill. Because we special-case github.com repos and don't yet deal with renames/deletions, there will be no bad effects.

fix #10121