default config to disablePublicRepoRedirects: true (#10128)
This disables "federation" and makes it require explicit admin action to enable public repo redirects.

This means that dev and Sourcegraph Server default config differs from the Data Center default config (the latter enables public repo redirects by default). This is not ideal, but we will address this in the docs.