Fix bug where goreman would not restart processes a second time using Go 1.10
We recently upgraded to Go 1.10 on CI. This unfortunately introduced
a regression in which goreman would fail to restart our processes more than
once. This was due to our goreman code being extremely complex, and having
a large amount of sync primitives sprinkled everywhere in a psychopathic
manner.

Reproducing the issue was as simple as:

1. Start a Sourcegraph Server Docker image built with Go 1.10.
2. Change anything in the site config that causes the "Restart" button to appear.
3. Clicking restart works the first time around.
4. Repeat steps 2 & 3. Server does not restart even though goreman is asked to via `net/rpc` call.

When using a Docker image built with Go 1.9, the above issue did not occur.

The main issue I found is that in `(Goreman).RestartAll`:

```Go
	// Prevent the server from shutting down (which it does when all processes are stopped).
	wg.Add(1)
	defer wg.Done()

	stopProcs(false)
	return startProcs()
```

`wg.Done()` being inside of a `defer` meant that `startProcs` cannot ever
return due to the processes exiting, as it should be doing. This is because it
relies on `wg.Wait()` to detect when the processes have all exited, but we've
artificially incremented it and will only decrement it once `startProcs`
returns (a catch-22).

But, why then did this code work in Go 1.9? I'm not entirely sure. I couldn't
find the answer after many hours of looking at this code. What I know is that:

1. The old code does work in Go 1.9, just not Go 1.10.
2. With this change, the code does work in Go 1.10.
3. Because our RPC server calls `rpc.ServeConn(client)` not in a goroutine, it
   can only serve one goroutine at a time. So when `startProcs()` blocked in Go
   1.10, the RPC server would lockup (and subsequent requests would be pending
   us handling the other one, which is why only the first request worked).

Fixes #10146