lsp-proxy: Always free associated VFS resources (#10170)
* lsp-proxy: Close rootFS in all cases init can fail

This is one of the ways we could not call close on rootFS, which leads to
unbounded disk growth of the archive cache.

* lsp-proxy: Always close rootFS when server is removed

Previously we would only free up rootFS when we call shutdownAndExit. However,
shutdownAndExit only gets called in the happy path of the proxy deciding we no
longer need the server. The other common cases (such as a server restarting)
would not be covered. Instead we use a far more reliable approach, which is
free up the resources once the connection dies.