trigger clone/index of repos after enabling (#10172)
This fixes an issue where if you visit a repo at http://sourcegraph.example.com/github.com/my/repo and click Add (if not added) or Enable (if disabled), the page would reload and immediately show 404 revision not found. See the HACK comment for why this occurs.

Previously, the HACK fully worked (to avoid the 404 revision not found). However, for some reason it no longer works reliably. I noticed that nothing in the SetRepositoryEnabled call actually triggers a clone, so no forward progress was actually being made during the 1500ms delay. This commit makes it so it kicks off cloning before the 1500ms delay starts, so it's hopefully cloning-in-progress by the end of the 1500ms delay.

Fixing for real this is known tech debt at #9801.