graphqlbackend: be robust against generating invalid commit URLs like "repo@/-/blob/..."
When `r.inputRev == nil` and `r.oid == ""`, the commit URL should NOT contain an `@rev`. Before this commit, the URLs being generated were of the form `/my/repo@/-/blob/...`, which is invalid. This occurs for file match results' commit resolvers returned by Zoekt, which always operates on the (last-indexed commit of) default branch.

Fixes #10272 for Zoekt results