search: show search suggestions for files whose paths match the query, even in subdirs
Previously a query like `file:a b` or `repo:a file:b c` would NOT show files whose paths matched. This meant that the DirectoryPage in a repo subdir would not let you jump to a file just by typing in its (fuzzy) path. This commit makes it so that's possible.