symbols: log/wrap this loop iter's parseErr, not the outer err (for which we don't hold the lock) (#10270)
I think it was a mistake that it logged `err` not `parseErr`. I was seeing log lines like:

```
15:16:05           symbols | t=2018-03-17T15:16:05-0700 lvl=eror msg="Error parsing symbols." repo=chromium.googlesource.com/chromium/src commitID=c98d55f133677d5281011be711833a488825ff5f path=components/multidevice/service/fake_device_sync.cc dataSize=953 error=nil
```

which were not helpful since all it says is `error=nil`. This occurred because `errors.Wrap(nil, ...)` returns `nil`, so `err` was stuck at `nil`.

Also I was seeing log lines like the following that should only happen if