frontend: Support Data Center for instrumentation page (#10302)
We can't hardcode the list of service hosts like we do for Server. So we instead query the kubernetes API similarly to how Prometheus does it. Except it we don't use the more advanced caching/discovery (since this page will rarely be queried), and we tailor it to our environment.