searcher: Literal Substring optimization (#10301)
The regular expression engine currently uses an optimization where it uses
`bytes.Index` if the regular expression has a literal prefix. This greatly
speeds up matching. However, if the string does not start with a literal
prefix we fallback to using slower matching for the whole file. Given that
most files do not have a match, we can ignore files that do not contain any
literal (not just prefix). So this commit introduces a function to extract the
longest literal from a regular expression, and uses that to filter files
before running the regular expression on it.