use latest go-langserver (#10356)
Previously we were on the no-godef branch. @slimsag thought that it would be OK to get back on master because godef usage is off by default, but it just needed to be tested. I've tested it and it seems to work (and all tests still pass). So I think we are good to go.

Similar to #9638 (that attempt was aborted)

The motivation is to make it easier to track go-langserver upstream improvements, including a soon-to-merge "find Go impls and interfaces" guru feature.

No changelog entry needed.

dep ensure -update github.com/sourcegraph/go-langserver