UserSettingsIntegrationsPage: update/tighten wording, show code intel alert (#10413)
Tightens up the wording on this page and shows a code intel alert if code intel is not enabled. Now that the extension supports Phab as well and has a more sophisticated menu itself (to show status of GitHub/GHE/etc.), we can remove some of the functionality on this page.