xlang: Server side of hover federation (#10433)
This is the server side (sourcegraph.com) of federating hover. It receives the
symbols which we attempt to resolve into a local definition we can hover
over.

An alternate design would involve just exposing `xdefQuery`. However, this
complicates the logic for the client significantly since it would then need to
send a single shot LSP query for each xdefQuery result. This pattern can also
easily be applied for definition and references.