Return on timeout instead of breaking
If breaking on timeout, the parseRequests channel will not be drained.
So fetchRepositoryArchive will block on writing to the channel as soon as the
buffer size is exceeded, which means it never gets to close the errChan,
which means this function will hang on reading from errChan.

There is no reason to not return as partial results here are of limited use
anyway because of caching and filtering.