generate external service links in new package externallink (#10505)
Previously the logic for generating links to external services (such as "view this repo/file/dir/commit on $CODEHOST") was scattered among many places.

For example, commits were passed through a redirect HTTP handler on our end, which meant that users could not easily right-click the link and copy the URL. GitHub Enterprise was special-cased on the frontend, which violated separation of concerns. And the graphqlbackend and the React component GoToCodeHost had different handling of the various types of services we supported (e.g., checking for URL prefixes to determine if the thing lived on GitHub in two different ways).

This standardizes all that in one subpackage of graphqlbackend, package externallink.