cmd/indexer: wait for frontend to start up before running workers
Since the indexer now relies on the frontend, and startup order is generally
never guaranteed, we get errors on Server startup like:

```
[...] msg="Could not list unindexed repositories" err="Post: Post http://127.0.0.1:3090/.api/graphql?ListRepos: dial tcp 127.0.0.1:3090: connect: connection refused"
```

Due to the fact that the frontend has not started yet we're trying to make
GraphQL requests to it. This is essentially just logspam, i.e. it does not
cause any real user issues today.

To fix it, we just wait for the frontend to be available as we do in other
services.

Fixes #10551