pkg/conf: apply config defaults in the right location (parseConfig)
Prior to this change we applied this 'convenience' default in the wrong
location. Callers of `parseConfig` would get different defaults than callers
of `initConfig`.

This was a pain to track down, but clearly this code was obviously added to
the wrong location by whoever added it. This caused an issue where, when we
compare diffs of configs (our diffing algorithm is simple/naive), we would
find one `experimentalFeatures` value of `nil` and another of an empty
`&schema.ExperimentalFeatures{}` and because they are different we would mark
the server as needing a restart.

Fixes #10570