BlobPanel: show error if url asks for references on bad location (#10656)
If the URL hash specifically asks for references on a location, the panel will always show the references tab. If the location is bad (e.g., not an ident), the error from the lang server will be displayed. If the user clicks on another tab (e.g., "File history", which always shows), then the bad tab will disappear.

fix #10542