graphqlbackend: Retry failed requests to searcher (#10612)
* graphqlbackend: Factor out request logic in textSearch

Will be using this to make retry logic clearer.

* graphqlbackend: Retry failed requests to searcher

* pkg/endpoint: Copy consistenthash implementation

We want to modify it for our needs.

* pkg/endpoint: Add a exclude list for Get

This will allow us to pick more appropriate endpoints retries. The algorithm
will pick the same endpoint as if all the endpoints in exclude where not part
of the map. This is especially useful since during deploys we will prefer the
older host until the new one is available.

* graphqlbackend: Do not pick the same host on retry for textSearch

See previous commit for why this is a good idea.

* increment attempt

* trace print if searcherURL lookup fails